# opt: add JoinLimit session variable