--- a/pkg/sql/exec_util.go
+++ b/pkg/sql/exec_util.go
@@ -1660,8 +1660,8 @@ func (m *sessionDataMutator) SetZigzagJoinEnabled(val bool) {
 	m.data.ZigzagJoinEnabled = val
 }
 
-func (m *sessionDataMutator) SetReorderJoins(val bool) {
-	m.data.ReorderJoins = val
+func (m *sessionDataMutator) SetReorderJoinsLimit(val int) {
+	m.data.ReorderJoinsLimit = val
 }
 
 func (m *sessionDataMutator) SetVectorize(val sessiondata.VectorizeExecMode) {