--- a/pkg/sql/logictest/testdata/logic_test/pg_catalog
+++ b/pkg/sql/logictest/testdata/logic_test/pg_catalog
@@ -1341,7 +1341,7 @@ default_transaction_read_only      off           NULL      NULL        NULL     
 distsql                            off           NULL      NULL        NULL        string
 experimental_enable_zigzag_join    off           NULL      NULL        NULL        string
 experimental_force_split_at        off           NULL      NULL        NULL        string
-experimental_reorder_joins         off           NULL      NULL        NULL        string
+experimental_reorder_joins_limit   0             NULL      NULL        NULL        string
 experimental_serial_normalization  rowid         NULL      NULL        NULL        string
 experimental_vectorize             off           NULL      NULL        NULL        string
 extra_float_digits                 0             NULL      NULL        NULL        string
@@ -1386,7 +1386,7 @@ default_transaction_read_only      off           NULL  user     NULL      off   
 distsql                            off           NULL  user     NULL      off           off
 experimental_enable_zigzag_join    off           NULL  user     NULL      off           off
 experimental_force_split_at        off           NULL  user     NULL      off           off
-experimental_reorder_joins         off           NULL  user     NULL      off           off
+experimental_reorder_joins_limit   0             NULL  user     NULL      0             0
 experimental_serial_normalization  rowid         NULL  user     NULL      rowid         rowid
 experimental_vectorize             off           NULL  user     NULL      off           off
 extra_float_digits                 0             NULL  user     NULL      0             2
@@ -1428,7 +1428,7 @@ distsql                            NULL    NULL     NULL     NULL        NULL
 experimental_enable_zigzag_join    NULL    NULL     NULL     NULL        NULL
 experimental_force_split_at        NULL    NULL     NULL     NULL        NULL
 experimental_optimizer_mutations   NULL    NULL     NULL     NULL        NULL
-experimental_reorder_joins         NULL    NULL     NULL     NULL        NULL
+experimental_reorder_joins_limit   NULL    NULL     NULL     NULL        NULL
 experimental_serial_normalization  NULL    NULL     NULL     NULL        NULL
 experimental_vectorize             NULL    NULL     NULL     NULL        NULL
 extra_float_digits                 NULL    NULL     NULL     NULL        NULL