--- a/pkg/sql/logictest/testdata/logic_test/show_source
+++ b/pkg/sql/logictest/testdata/logic_test/show_source
@@ -38,7 +38,7 @@ default_transaction_read_only      off
 distsql                            off
 experimental_enable_zigzag_join    off
 experimental_force_split_at        off
-experimental_reorder_joins         off
+experimental_reorder_joins_limit   0
 experimental_serial_normalization  rowid
 experimental_vectorize             off
 extra_float_digits                 0