--- a/pkg/sql/opt/exec/execbuilder/testdata/join_order
+++ b/pkg/sql/opt/exec/execbuilder/testdata/join_order
@@ -56,7 +56,7 @@ render               ·                   ·                (a, b, c, d, b, x, c
 ·                    spans               /1-/1/#          ·                         ·
 
 statement ok
-SET experimental_reorder_joins = true
+SET experimental_reorder_joins_limit = 3
 
 query TTTTT
 EXPLAIN (VERBOSE) SELECT * FROM abc, bx, cy WHERE a = 1 AND abc.b = bx.b AND abc.c = cy.c