--- a/pkg/sql/opt/xform/testdata/rules/join_order
+++ b/pkg/sql/opt/xform/testdata/rules/join_order
@@ -50,7 +50,7 @@ TABLE abc
  └── INDEX primary
       └── a int not null
 
-opt reorder-joins
+opt join-limit=3
 SELECT * FROM abc, bx, cy WHERE a = 1 AND abc.b = bx.b AND abc.c = cy.c
 ----
 inner-join (lookup bx)
@@ -74,7 +74,7 @@ inner-join (lookup bx)
  │    └── filters (true)
  └── filters (true)
 
-opt reorder-joins
+opt join-limit=3
 SELECT * FROM bx, abc, cy WHERE a = 1 AND abc.b = bx.b AND abc.c = cy.c
 ----
 inner-join (lookup bx)
@@ -98,7 +98,7 @@ inner-join (lookup bx)
  │    └── filters (true)
  └── filters (true)
 
-opt reorder-joins
+opt join-limit=3
 SELECT * FROM bx, cy, abc WHERE a = 1 AND abc.b = bx.b AND abc.c = cy.c
 ----
 inner-join (lookup bx)
@@ -122,7 +122,31 @@ inner-join (lookup bx)
  │    └── filters (true)
  └── filters (true)
 
-memo reorder-joins
+opt join-limit=2 expect-not=AssociateJoin
+SELECT * FROM bx, cy, abc WHERE a = 1 AND abc.b = bx.b AND abc.c = cy.c
+----
+inner-join (lookup bx)
+ ├── columns: b:1(int!null) x:2(int) c:3(int!null) y:4(int) a:5(int!null) b:6(int!null) c:7(int!null) d:8(int)
+ ├── key columns: [6] = [1]
+ ├── cardinality: [0 - 1]
+ ├── key: ()
+ ├── fd: ()-->(1-8)
+ ├── inner-join (lookup cy)
+ │    ├── columns: cy.c:3(int!null) y:4(int) a:5(int!null) abc.b:6(int) abc.c:7(int!null) d:8(int)
+ │    ├── key columns: [7] = [3]
+ │    ├── cardinality: [0 - 1]
+ │    ├── key: ()
+ │    ├── fd: ()-->(3-8)
+ │    ├── scan abc
+ │    │    ├── columns: a:5(int!null) abc.b:6(int) abc.c:7(int) d:8(int)
+ │    │    ├── constraint: /5: [/1 - /1]
+ │    │    ├── cardinality: [0 - 1]
+ │    │    ├── key: ()
+ │    │    └── fd: ()-->(5-8)
+ │    └── filters (true)
+ └── filters (true)
+
+memo join-limit=3
 SELECT * FROM bx, cy, abc WHERE a = 1 AND abc.b = bx.b AND abc.c = cy.c
 ----
 memo (optimized, ~17KB, required=[presentation: b:1,x:2,c:3,y:4,a:5,b:6,c:7,d:8])
@@ -179,7 +203,7 @@ memo (optimized, ~17KB, required=[presentation: b:1,x:2,c:3,y:4,a:5,b:6,c:7,d:8]
  ├── G21: (variable a)
  └── G22: (const 1)
 
-opt reorder-joins
+opt join-limit=4
 SELECT * FROM bx, cy, dz, abc WHERE a = 1
 ----
 inner-join
@@ -214,6 +238,73 @@ inner-join
  │    ├── columns: bx.b:1(int!null) x:2(int)
  │    ├── key: (1)
  │    └── fd: (1)-->(2)
+ └── filters (true)
+
+opt join-limit=3 format=show-all
+SELECT * FROM abc, bx, cy, dz WHERE a = 1
+----
+inner-join
+ ├── columns: a:1(int!null) b:2(int) c:3(int) d:4(int) b:5(int!null) x:6(int) c:7(int!null) y:8(int) d:9(int!null) z:10(int)
+ ├── stats: [rows=1e+09]
+ ├── cost: 32525668.7
+ ├── key: (5,7,9)
+ ├── fd: ()-->(1-4), (5)-->(6), (7)-->(8), (9)-->(10)
+ ├── prune: (2-10)
+ ├── interesting orderings: (+7) (+9) (+5) (+1)
+ ├── inner-join
+ │    ├── columns: t.public.bx.b:5(int!null) t.public.bx.x:6(int) t.public.cy.c:7(int!null) t.public.cy.y:8(int) t.public.dz.d:9(int!null) t.public.dz.z:10(int)
+ │    ├── stats: [rows=1e+09]
+ │    ├── cost: 10025667.5
+ │    ├── key: (5,7,9)
+ │    ├── fd: (5)-->(6), (7)-->(8), (9)-->(10)
+ │    ├── prune: (5-10)
+ │    ├── interesting orderings: (+7) (+9) (+5)
+ │    ├── join-size: 3
+ │    ├── inner-join
+ │    │    ├── columns: t.public.cy.c:7(int!null) t.public.cy.y:8(int) t.public.dz.d:9(int!null) t.public.dz.z:10(int)
+ │    │    ├── stats: [rows=1000000]
+ │    │    ├── cost: 12110.03
+ │    │    ├── key: (7,9)
+ │    │    ├── fd: (7)-->(8), (9)-->(10)
+ │    │    ├── prune: (7-10)
+ │    │    ├── interesting orderings: (+7) (+9)
+ │    │    ├── join-size: 2
+ │    │    ├── scan t.public.cy
+ │    │    │    ├── columns: t.public.cy.c:7(int!null) t.public.cy.y:8(int)
+ │    │    │    ├── stats: [rows=1000]
+ │    │    │    ├── cost: 1040.01
+ │    │    │    ├── key: (7)
+ │    │    │    ├── fd: (7)-->(8)
+ │    │    │    ├── prune: (7,8)
+ │    │    │    └── interesting orderings: (+7)
+ │    │    ├── scan t.public.dz
+ │    │    │    ├── columns: t.public.dz.d:9(int!null) t.public.dz.z:10(int)
+ │    │    │    ├── stats: [rows=1000]
+ │    │    │    ├── cost: 1040.01
+ │    │    │    ├── key: (9)
+ │    │    │    ├── fd: (9)-->(10)
+ │    │    │    ├── prune: (9,10)
+ │    │    │    └── interesting orderings: (+9)
+ │    │    └── filters (true)
+ │    ├── scan t.public.bx
+ │    │    ├── columns: t.public.bx.b:5(int!null) t.public.bx.x:6(int)
+ │    │    ├── stats: [rows=1000]
+ │    │    ├── cost: 1040.01
+ │    │    ├── key: (5)
+ │    │    ├── fd: (5)-->(6)
+ │    │    ├── prune: (5,6)
+ │    │    └── interesting orderings: (+5)
+ │    └── filters (true)
+ ├── scan t.public.abc
+ │    ├── columns: t.public.abc.a:1(int!null) t.public.abc.b:2(int) t.public.abc.c:3(int) t.public.abc.d:4(int)
+ │    ├── constraint: /1: [/1 - /1]
+ │    ├── cardinality: [0 - 1]
+ │    ├── stats: [rows=1, distinct(1)=1, null(1)=0]
+ │    ├── cost: 1.09
+ │    ├── key: ()
+ │    ├── fd: ()-->(1-4)
+ │    ├── prune: (2-4)
+ │    └── interesting orderings: (+1)
  └── filters (true)
 
 # Note the difference in memo size for with and without reorder-joins, for only four tables.
@@ -261,7 +352,7 @@ memo (optimized, ~11KB, required=[presentation: b:1,x:2,c:3,y:4,d:5,z:6,a:7,b:8,
  ├── G12: (variable a)
  └── G13: (const 1)
 
-memo reorder-joins
+memo join-limit=4
 SELECT * FROM bx, cy, dz, abc WHERE a = 1
 ----
 memo (optimized, ~25KB, required=[presentation: b:1,x:2,c:3,y:4,d:5,z:6,a:7,b:8,c:9,d:10])