--- a/pkg/sql/sessiondata/session_data.go
+++ b/pkg/sql/sessiondata/session_data.go
@@ -65,9 +65,9 @@ type SessionData struct {
 	// ZigzagJoinEnabled indicates whether the optimizer should try and plan a
 	// zigzag join.
 	ZigzagJoinEnabled bool
-	// ReorderJoins indicates whether the optimizer should try to reorder the
-	// joins in a query.
-	ReorderJoins bool
+	// ReorderJoinsLimit indicates the number of joins at which the optimizer should
+	// stop attempting to reorder.
+	ReorderJoinsLimit int
 	// SequenceState gives access to the SQL sequences that have been manipulated
 	// by the session.
 	SequenceState *SequenceState