This commit introduces the JoinLimit session variable, which enables a
transformation-based search of potential join orderings when a query
contains no more than a specified number of joins. I think we could
theoretically safely set this value to a default of 4 or 5 today, and I
suspect without too much trouble we can speed it up enough to bump it up
to 6 or 7.

This is sort of a crude heuristic: we count the number of joins
*globally*, even though those might not all be part of the same join
expression.

I briefly tried to compute the size of the biggest join tree via a
logical property, but that wasn't great, since we'd want it to be a Rule
property, and the Shared properties don't have those. In the end I
decided this was probably simplest (we just walk the tree after build).

This method does have some perf impact:

Before:
```
BenchmarkPhases/kv-read/Explore-8         	  100000	     12449 ns/op
BenchmarkPhases/kv-read-no-prep/Explore-8 	   30000	     40384 ns/op
BenchmarkPhases/kv-read-const/Explore-8   	10000000	       120 ns/op
BenchmarkPhases/tpcc-new-order/Explore-8  	   30000	     41891 ns/op
BenchmarkPhases/tpcc-delivery/Explore-8   	   50000	     26807 ns/op
BenchmarkPhases/tpcc-stock-level/Explore-8     	   10000	    134082 ns/op
```

After:
```
BenchmarkPhases/kv-read/Explore-8         	  100000	     12216 ns/op
BenchmarkPhases/kv-read-no-prep/Explore-8 	   30000	     41637 ns/op
BenchmarkPhases/kv-read-const/Explore-8   	10000000	       123 ns/op
BenchmarkPhases/tpcc-new-order/Explore-8  	   30000	     42511 ns/op
BenchmarkPhases/tpcc-delivery/Explore-8   	   50000	     27790 ns/op
BenchmarkPhases/tpcc-stock-level/Explore-8         10000	    137069 ns/op
```

Alternatively we could increment the counter in the memo during
the build phase, which might reduce the perf impact.

Next steps also include adding reordering rules for outer joins.

Planning time for chain queries:
<img width="658" alt="image" src="https://user-images.githubusercontent.com/409075/51936630-a2094200-23d6-11e9-847c-5b12ab976fff.png">

Release note (sql change): replaced experimental_reorder_joins session
variable with experimental_join_limit, which will reorder joins in
queries with fewer than the specified number of joins.