@rytaft 2019-02-04
___
*[pkg/sql/opt/xform/testdata/rules/join_order, line 183 at r2](https://reviewable.io/reviews/cockroachdb/cockroach/34375#-LXtDScODK-ScBZfeJiC:-LXtDScODK-ScBZfeJiD:bklvqcw) ([raw file](https://github.com/cockroachdb/cockroach/blob/d5f3bee631a7dbfb07b9da165ad0b25eca4cf69a/pkg/sql/opt/xform/testdata/rules/join_order#L183)):*
> ```
> 
> opt join-limit=4
> SELECT * FROM bx, cy, dz, abc WHERE a = 1
> ```

just realized that this produces the same plan as the one below. Maybe you can inject some stats so they produce different plans?


<!-- Sent from Reviewable.io -->